Decouple component, add config
